+++ client/emajParallelRollback.pl
@@ -52,9 +52,10 @@ my $password = undef;                     # -W user password
 my $groups = undef;                       # -g E-maj group names (mandatory)
 my $mark = undef;                         # -m E-maj mark name to rollback to (mandatory)
 my $nbSession = 1;                        # -s number of parallel E-maj sessions to use for rollback (default=1)
-my $verbose = 0;                          # -v flag for verbose mode
+my $comment = '';                         # -c rollback comment
 my $isLogged = 'false';                   # -l flag for logged rollback mode
 my $isAlterGroupAllowed = 'false';        # -a flag to allow the rollback to reach a mark set before alter group operations
+my $verbose = 0;                          # -v flag for verbose mode
 
 my  $conn_string = '';
 my  $multiGroup = 'false';
@@ -70,6 +71,7 @@ GetOptions(
   "W=s" => \$password,
 #  other parameters
   "a"   => sub { $isAlterGroupAllowed = 'true'; },
+  "c:s" => \$comment,
   "g=s" => \&check_opt_groups,
   "l"   => sub {$isLogged = 'true'; $msgRlbk = 'Logged rollback'; },
   "m:s" => \$mark,
@@ -132,7 +134,8 @@ for (my $i = 1 ; $i <= $nbSession; $i++) {
 # Call _rlbk_init() on first session
 # This checks the groups and mark, and prepares the parallel rollback by creating well balanced sessions
 print (strftime('%d/%m/%Y - %H:%M:%S',localtime)." Call _rlbk_init() for groups $groups and mark $mark....\n") if ($verbose);
-$stmt[1] = $dbh[1]->prepare("SELECT emaj._rlbk_init(array[$groups], " . $dbh[1]->quote($mark) . ", $isLogged, $nbSession, $multiGroup, $isAlterGroupAllowed)");
+$stmt[1] = $dbh[1]->prepare("SELECT emaj._rlbk_init(array[$groups], " . $dbh[1]->quote($mark) . ", $isLogged, $nbSession, $multiGroup, " .
+                                                   "$isAlterGroupAllowed, " . $dbh[1]->quote($comment) . ")");
 $stmt[1]->execute()
   or die("Calling the _rlbk_init() function failed.\n$DBI::errstr\n");
 my $rlbkId = $stmt[1]->fetchrow_array();
@@ -208,8 +211,10 @@ for (my $i = 1 ; $i <= $nbSession; $i++) {
 }
 #clean up on error
 END {
-  for (my $i = 1 ; $i <= $nbSession; $i++) {
-	$dbh[$i]->disconnect if ( ($dbh[$i]) && ($dbh[$i]->{Active}) )
+  if (defined($nbSession)) {
+    for (my $i = 1 ; $i <= $nbSession; $i++) {
+      $dbh[$i]->disconnect if ( ($dbh[$i]) && ($dbh[$i]->{Active}) )
+    }
   }
 }
 
@@ -232,6 +237,7 @@ Options:
 
   -l          logged rollback mode (i.e. 'rollbackable' rollback)
   -a          flag to allow rollback to reach a mark set before alter group operations
+  -c          comment to describe the rollback operation
   -v          verbose mode; writes more information about the processing
   --help      shows this help, then exit
   --version   outputs version information, then exit
@@ -244,12 +250,12 @@ Connection options:
   -W,         password associated to the user, if needed
   
 Examples:
-  $PROGRAM -g myGroup1 -m myMark -s 3
-              performs a parallel rollback of table group myGroup1 to mark
-              myMark using 3 parallel sessions.
-  $PROGRAM -h localhost -p 5432 -d myDb -U emajadm -l -g \"myGroup1,myGroup2\" -m myMark -s 5 -v
-              lets the role emajadm perform a parallel logged rollback of 2 table
-              groups to mark myMark using 5 parallel sessions, in verbose mode.
+  $PROGRAM -g myGroup1 -m myMark -s 3 -c \"Revert aborted ABC chain\"
+              performs a parallel rollback of the tables group myGroup1 to the mark
+              myMark, using 3 parallel sessions, with a comment.
+  $PROGRAM -h localhost -p 5432 -d myDb -U emajadmin -l -g \"myGroup1,myGroup2\" -m myMark -s 5 -v
+              lets the role emajadmin perform a parallel logged rollback of 2 table
+              groups to the mark myMark using 5 parallel sessions, in verbose mode.
 };
 }